Split minimal FSDP runtime modules
