+++ megatron/core/distributed/fsdp/src/megatron_fsdp/experimental/__init__.py
@@ -15,8 +15,9 @@
 """Experimental Megatron-FSDP implementation."""
 
 from .dbuffer import DBuffer
-from .fully_shard import FsdpModule, ParameterGroup, Placements, fully_shard
-from .placement import Flat, Partial, Placement, Replicate
+from .fully_shard import FsdpModule, fully_shard
+from .parameter_group import ParameterGroup
+from .placement import Flat, Partial, Placement, Placements, Replicate
 
 __all__ = [
     "DBuffer",
